More complete ExceptionHandler.framework

+++ src/ExceptionHandling/ExceptionHandler.mm
@@ -3,6 +3,7 @@
 #include <sys/types.h>
 #include <signal.h>
 #include <unistd.h>
+#include <cstdlib>
 
 static void LocalExceptionHandler(NSException* e);
 
@@ -20,9 +21,9 @@ static void LocalExceptionHandler(NSException* e);
 	pthread_once(&once, []() {
 	
 		instance = [NSExceptionHandler new];
-		instance.delegate = NULL;
-		instance.exceptionHandlingMask = 0xffffffff;
-		instance.exceptionHangingMask = 0;
+		instance->_delegate = NULL;
+		instance->_handlingMask = 0xffffffff;
+		instance->_hangingMask = 0;
 		
 		NSSetUncaughtExceptionHandler(LocalExceptionHandler);
 	});
@@ -37,11 +38,32 @@ void LocalExceptionHandler(NSException* e)
 	NSExceptionHandler* handler = [NSExceptionHandler defaultExceptionHandler];
 	unsigned int mask = handler.exceptionHandlingMask;
 	unsigned int hangmask = handler.exceptionHangingMask;
-	
-	if (mask & NSLogUncaughtExceptionMask)
-		NSLog(@"Uncaught exception: %@", e);
-	if (hangmask & NSHangOnUncaughtExceptionMask)
-		kill(getpid(), SIGTRAP);
-	if (mask & NSHandleUncaughtExceptionMask)
-		abort();
+	id delegate = handler.delegate;
+
+	if (delegate != NULL)
+	{
+		bool handle, log;
+
+		handle = [delegate exceptionHandler: [NSExceptionHandler defaultExceptionHandler]
+		              shouldHandleException: e
+                                       mask: mask];
+
+		log = [delegate exceptionHandler: [NSExceptionHandler defaultExceptionHandler]
+		              shouldLogException: e
+                                    mask: mask];
+
+		if (log)
+			NSLog(@"Uncaught exception: %@", e);
+		if (handle)
+			abort();
+	}
+	else
+	{
+		if (mask & NSLogUncaughtExceptionMask)
+			NSLog(@"Uncaught exception: %@", e);
+		if (hangmask & NSHangOnUncaughtExceptionMask)
+			kill(getpid(), SIGTRAP);
+		if (mask & NSHandleUncaughtExceptionMask)
+			abort();
+	}
 }
